requirements.txt
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
python-dotenv
azure-communication-email